-updated gamecube dumper again, enabled 32k align by default now, compression without 32k still doesnt work properly -fixed stupid bug which may deletes all your games when deleting one game (thanks Miigotu) -reverted video change of r154 for now

+++ source/loader/wbfs_ext.c
@@ -163,10 +163,12 @@ s32 WBFS_Ext_RemoveGame(u8 *discid, char *gamepath)
 		if (ent->d_name[0] == '.') continue;
 
 		snprintf(file, sizeof(file), "%s/%s", folder, ent->d_name);
-		remove(file);		
+		if(discid != NULL && strstr(file, (char*)discid) != NULL)
+			remove(file);
 	}
 	closedir(dir_iter);
-	unlink(folder);
+	if(strlen(folder) > 11)
+		unlink(folder);
 	return 0;
 }
